% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{annotate_gene}
\alias{annotate_gene}
\title{Gene annotation}
\usage{
annotate_gene(upstream, downstream)
}
\arguments{
\item{upstream}{numeric value of upsteam length from the start site of gene. Default is \code{10}.
Unit is "kb".}

\item{downstream}{numeric value of downstream length from the end site of gene. Default is \code{1.5}.
Unit is "kb".}
}
\value{
A data frame of gene annotations: gene name, chromosome, start site of annotation,
 end site of annotation, strand, and gene window size in kb.
}
\description{
We annotate each gene by defining a gene window with 10kb upstream of start site
 and 1.5kb downstream of end site, by default. One column of gene window length in kb is appended
 for future use in gene-wise LD pruning.
}
\examples{
\dontrun{
 anno_dictionary <- annotate_gene(upstream = 10, downstream = 1.5)
 write.table(anno_dictionary, file = "annotation_dictionary.txt", quote = FALSE,
             sep = "\\t", row.names = FALSE, col.names = FALSE)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
