\name{binary_search}
\alias{binary_search}
\title{
Find intervals by binary search
}
\description{
Find intervals by binary search
}
\usage{
binary_search(breaks, search, left_index = TRUE)
}
\arguments{

  \item{breaks}{a non-decreasing integer vector}
  \item{search}{a iteger vector}
  \item{left_index}{whether to use the index of left break or right break}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
