% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{BRef}
\alias{BRef}
\alias{BRef.tpm}
\title{Reference profiles from circulating immune cells.}
\format{A list of 3 elements: \describe{ \item{$refProfiles,
  $refProfiles.var}{Matrices (nGenes x nRefCells) of the normalized gene
  expression from the reference cells and the variability of this gene
  expression for each gene and each cell type} \item{$sigGenes}{A list of 100
  signature genes used to deconvolve the cell proportions} }}
\source{
\enumerate{
 \item \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-GEOD-64655},
 \item \url{https://www.ebi.ac.uk/arrayexpress/experiments/E-GEOD-60424/},
 \item \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE51984}
 }
}
\usage{
BRef

BRef.tpm
}
\description{
A dataset containing the reference profiles obtained from immune cell
samples of \emph{B-}, \emph{NK-}, \emph{T-cells}, \emph{Monocytes}
and \emph{Neutrophils} purified from PBMC or whole blood.
}
\details{
The original samples were obtained from healthy donors and donors after
influenza vaccination or with diabetes, sepsis or multiple sclerosis.
}
\section{Similar datasets}{

 \code{BRef} (main reference profiles, using data from sources 1-3 below)


\code{BRef.tpm} (reference profiles based on same data as \code{BRef},
 but given in TPM counts instead of raw counts)
}

\keyword{datasets}
