% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{calculate_POET_sw}
\alias{calculate_POET_sw}
\title{POET estimator and sliding window approach}
\usage{
calculate_POET_sw(genotype, gene_pval, gene.loc, chr, type, inter_dir,
                   X_super = NULL, sliding_size = 10, Cmin = 1.5, sliding_step = 1)
}
\arguments{
\item{genotype}{a genotype matrix}

\item{gene_pval}{a list of GWAS summary statistics for genes}

\item{gene.loc}{a data frame of gene dictionary, only including genes of interest in each transcriptomic profile
(e.g., top 8000 highly variable genes from scRNA-seq or genes with great specificity score from GTEx v8).
This can be returned from \code{process_scRNA}.}

\item{chr}{numeric value of chromosome number}

\item{type}{type of variants included in POET estimation. This should be either 'prunein' or 'joint'.}

\item{inter_dir}{directory of intermediate output files}

\item{X_super}{a matrix of aggregated pseudo-SNP. Default is \code{NULL}.}

\item{sliding_size}{size of sliding window. Default is \code{10}.}

\item{Cmin}{a user-specified minimum constant of thresholding funciton, determining the POET shrinkage.
Default is \code{1.5}.}

\item{sliding_step}{the moving step of sliding-window. Default is \code{1}.}
}
\value{
Intermediate stored files including
 \item{.rda}{POET shrinkage correlation estimator}
 \item{.txt}{A list of genes residing in the same sliding window}
}
\description{
We adopt the POET estimator to obtain a well-conditioned covariance matrix
 via sparse shrinkage. We implement a sliding-window approach to only compute
 correlations for pairs of genes within a certain distance from each.
}
\examples{
\dontrun{
inter_dir <- file.path("/path/to/intermediates", trait)
calculate_POET_sw(genotype = X_ref.Demo, gene_pval = prunein_gene_pval.Demo,
                  gene.loc = gtex.loc,
                  chr = chr, type = "POET", inter_dir = inter_dir)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
