% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{ci_prop}
\alias{ci_prop}
\title{Confidence interval for proportion $\pi$}
\usage{
ci_prop(p, n, conf.level = 0.95, method = "wald", bounds = "two-sided")
}
\arguments{
\item{bounds}{option to choose two-sided confidence intervals, lower or upper bounds. Defaults to \code{two-sided}}

\item{p:}{sample proportion}

\item{n:}{sample size}

\item{conf.level:}{confidence level between 0 and 1. Defaults to .95}

\item{method:}{the method chosen for computing the confidence interval
for proportions. Defaults to \code{wald}}
}
\description{
Confidence intervals for the population proportion $\pi$ for a large \emph{n} and
the population mean $\mu$ under certain conditions.
}
\details{
When \emph{n} is sufficiently large and \emph{p} is not too small, the sample proportion \emph{p} has
an approximately Gaussian distribution. In the long run, if repeated samples were taken,
the sample proportion would vary around the proportion $\pi$ in the target population.

There are three available methods for computing the binomial proportion confidence interval.
\itemize{
\item \strong{wald}: Relies on approximating the distribution of error about a binomially-distributed
observation. Typically used when \emph{n} is sufficiently large and \emph{p} is not too small.
\item \strong{exact}: This method uses the binomial distribution to calculate the coverage probability, which
is never less than $1- \alpha$. See \href{https://stats.stackexchange.com/questions/87775/clopper-pearson-for-non-mathematicians}{here} for a simple explanation
\item \strong{wilson}: Asymetric interval with continuity correction typically employed with small samples and skewed observations.
The observed coverage probability is closer to $1- \alpha$ than the Wald method.
}
}
\examples{
ci_prop(.45,100)

}
