% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combine_gene_corr.mat.Demo}
\alias{combine_gene_corr.mat.Demo}
\title{Pre-stored combine_gene_corr.mat.Demo data for demonstration purposes}
\format{Pre-computed by EPIC using pre-stored data \code{combine_POET_gene_pval.Demo}}
\usage{
combine_gene_corr.mat.Demo
}
\description{
Pre-stored combine_gene_corr.mat.Demo data for demonstration purposes
}
\keyword{datasets}
