% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combine_POET_gene_pval.Demo}
\alias{combine_POET_gene_pval.Demo}
\title{Pre-stored combine_POET_gene_pval.Demo data for demonstration purposes}
\format{Pre-computed by EPIC using pre-stored data \code{gwas.formatted.Demo} and \code{X_ref.Demo}}
\usage{
combine_POET_gene_pval.Demo
}
\description{
Pre-stored combine_POET_gene_pval.Demo data for demonstration purposes
}
\keyword{datasets}
