% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{divide_common_rare}
\alias{divide_common_rare}
\title{Divide SNPs into common and rare variants}
\usage{
divide_common_rare(gwas, snp_to_gene)
}
\arguments{
\item{gwas}{a data frame of formatted GWAS summary statistics}

\item{snp_to_gene}{A list of assigned SNPs for genes, returned from \code{map_snp_to_gene}.}

\item{maf_thres}{the MAF cutoff for common and rare variants. Default is \code{0.01}.}

\item{mac_lb}{the minor allele count (MAC) lower bound. A gene is excluded from the analysis if its MAC is smaller
than the MAC lower bound. Default is \code{20}.}
}
\value{
A list with components
 \item{common_snp_to_gene}{A list of common variants for genes}
 \item{rare_snp_to_gene}{A list of rare variants for genes}
}
\description{
By default, EPIC use MAF=1% as the cutoff for common and rare variants.
 For rare variants, the upper bound of inclusion is controlled by MAF
 while the lower bound is determined by MAC (Minor Allele Count).
 Rare variants with MAC less than 20 are removed from analysis.
}
\examples{
\dontrun{
snp_division_obj <- divide_common_rare(gwas = gwas.formatted, snp_to_gene = snp_to_gene)
common_snp_to_gene <- snp_division_obj$common_snp_to_gene
rare_snp_to_gene <- snp_division_obj$rare_snp_to_gene
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
