% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{package}
\name{EPIC.package}
\alias{EPIC.package}
\alias{EPIC.package-package}
\title{EPIC: a package to Estimate the Proportion of Immune and Cancer cells from
 tumor gene expression data.}
\description{
EPIC package provides the function and immune cell reference profiles to
estimate the proportion of immune and other cells from bulk gene expression
data.
}
\section{EPIC functions}{

\code{\link{EPIC}} is the main function to call to estimate the
 various cells proportions from a bulk sample.
}

\section{Included datasets}{

\code{\link{BRef}}, \code{\link{BRef.tpm}}: reference profiles from
   circulating immune cells.

 \code{\link{TRef.tpm}}: reference profiles from immune cells obtained
 from single cell data of tumor infiltrating cells from melanoma patients.

\code{\link{hoek_data}}: example dataset containing data from Hoek et al,
 2015, PLoS One.

\code{\link{mRNA_cell_default}}: values of mRNA per cell for the main cell
 types.
}

\section{Authors}{

Julien Racle <\email{julien.racle@unil.ch}> and David Gfeller <\email{
 david.gfeller@unil.ch}>.
}

