% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{format_gwas}
\alias{format_gwas}
\title{Formatting GWAS summary statistics}
\usage{
format_gwas(gwas, bfile_path, plink_path)
}
\arguments{
\item{gwas}{a data frame of GWAS summary statistics without formatting}

\item{bfile_path}{path to the \code{bim} file of reference panel}

\item{plink_path}{path to the PLINK toolkit}
}
\value{
A data frame of formatted GWAS summary statistics after quality control and allele alignment
}
\description{
Large-scale GWAS summary statistics from different consortiums
 are in diverse formats, due to the nature of phenotypes being studied (case-control
 traits v.s. quantitative traits) and various analytical tools. We recommend formatting
 summary statistics in post-GWAS analysis. We perform quality control and allele
 alignment with respect to the external reference panel.
}
\examples{
\dontrun{
bfile_path <- "/path/to/bfile"
plink_path <- "/path/to/PLINK"
gwas.formatted <- format_gwas(gwas, bfile_path, plink_path)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
