% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{get_burden}
\alias{get_burden}
\title{Burden test for rare variants}
\usage{
get_burden(genotype, gene_pval)
}
\arguments{
\item{genotype}{a genotype matrix}

\item{gene_pval}{a list of GWAS summary statistics for genes}
}
\value{
A list of gene-level testing for rare variants
}
\description{
We recover burden test statistics from GWAS summary statistics
 for the gene-level association testing of rare variants.
}
\examples{
\dontrun{
rare_gene_pval.Demo <- get_burden(genotype = X_ref.Demo, gene_pval = rare_gene_pval.Demo)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
