% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_combined}
\alias{get_combined}
\title{Combine gene-level testing for common and rare variants}
\usage{
get_combined(prunein_gene_pval, rare_gene_pval)
}
\arguments{
\item{prunein_gene_pval}{a list of gene-level testing for prune-in common variants}

\item{rare_gene_pval}{a list of gene-level testing for rare variants}
}
\value{
A list of combined gene-level testing
}
\description{
We combine gene-level chi-square association testing for common variants
 and burden test for rare variants. 1
}
\examples{
\dontrun{
combine_gene_pval <- get_combined(prunein_gene_pval, rare_gene_pval)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
