\name{get_peak_list}
\alias{get_peak_list}
\title{
Get a list of peak regions
}
\description{
Get a list of peak regions
}
\usage{
get_peak_list(mark)
}
\arguments{

  \item{mark}{mark type}

}
\details{
It works after \code{\link{chipseq_hooks}} is set
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
