% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex.Demo}
\alias{gtex.Demo}
\title{Pre-stored gtex.Demo data for demonstration purposes}
\format{Pre-processed by EPIC using GTEx v8 transcriptomic profiles}
\usage{
gtex.Demo
}
\description{
Pre-stored gtex.Demo data for demonstration purposes
}
\keyword{datasets}
