% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas.Demo}
\alias{gwas.Demo}
\title{Pre-stored gwas.Demo data for demonstration purposes}
\format{Provided by GLGC LDL summary statistics, Willer et al. 2013}
\usage{
gwas.Demo
}
\description{
Pre-stored gwas.Demo data for demonstration purposes
}
\keyword{datasets}
