% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas.formatted.Demo}
\alias{gwas.formatted.Demo}
\title{Pre-stored gwas.formatted.Demo data for demonstration purposes}
\format{Pre-computed by EPIC using pre-stored data \code{gwas.Demo}}
\usage{
gwas.formatted.Demo
}
\description{
Pre-stored gwas.formatted.Demo data for demonstration purposes
}
\keyword{datasets}
