% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{hoek_data}
\alias{hoek_data}
\title{Example dataset containing data from Hoek et al, 2015, PLoS One.}
\format{This is a list of 3 elements: \describe{
 \item{$rawCounts}{(matrix of 51574 genes x 2 donors) The raw read counts
   from the two donors. It has been obtained by mapping the original
   fastq files to \emph{hg19} genome with help of \emph{tophat} and
   \emph{htseq-count}.}
 \item{$cellFractions.obs}{(matrix of 2 donors x 6 cell types) The
   proportions of the different immune cells in the PBMC from the 2 donors,
   as measured by FACS by Hoek et al.}
 \item{$cellFractions.pred}{(matrix of 2 donors x 7 cell types) The
   proportions of the different immune cells and of a potential additional
   uncharacterized cell, as predicted by EPIC based on the rawCounts and
   the profiles \code{reference=BRef}.}
}}
\source{
The description of this data can be found here:
 \href{http://dx.doi.org/10.1371/journal.pone.0118528}{link to paper}
 and \href{https://www.ebi.ac.uk/arrayexpress/experiments/E-GEOD-64655}{link
 to data}.
}
\usage{
hoek_data
}
\description{
This dataset contains a subset of the full Hoek et al data. It contains only
the data from the two healthy donors PBMC before influenza vaccination.
}
\keyword{datasets}

