% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{influential_testing}
\alias{influential_testing}
\title{Identify a set of tissue- or cell-type-specific influential genes}
\usage{
influential_testing(gene_pval, gene_corr.mat, gene_expr, gene.loc, ct)
}
\arguments{
\item{gene_pval}{a list of genes with gene-level association testing}

\item{gene_corr.mat}{a matrix of gene-gene correlations}

\item{gene_expr}{a data frame of gene expressions by tissues or cell types. For scRNA-seq data,
this can be returned from \code{process_scRNA}.}

\item{gene.loc}{a data frame of gene dictionary, only including genes of interest in each transcriptomic profile
(e.g., top 8000 highly variable genes from scRNA-seq or genes with great specificity score from GTEx v8).
This can be returned from \code{process_scRNA}.}

\item{ct}{a character of prioritized tissue or cell type}
}
\value{
A list with components
 \item{influential.genes}{A list of tissue- or cell-type-specific influential genes}
 \item{p.dfbeta}{A ggplot object for DFBETA}
}
\description{
For a significantly enriched tissue or cell type,
 we further carry out a statistical influence test to identify a set of
 tissue- or cell-type-specific influential genes, using the DFBETAS statistics.
}
\examples{
\dontrun{
influential_test.gtex <- influential_testing(gene_pval = prunein_POET_gene_pval,
                                             gene_corr.mat = prunein_gene_corr.mat,
                                             gene_expr = gtex, gene.loc = gtex.loc,
                                             ct = "Liver")
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
