% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{LD_pruning}
\alias{LD_pruning}
\title{Perform LD pruning on common variants and import prune-in variants}
\usage{
LD_pruning(common_snp_to_gene, prunein_dir)
}
\arguments{
\item{common_snp_to_gene}{A list of assigned common SNPs for genes, returned from \code{divide_common_rare}.}

\item{prunein_dir}{the directory of LD pruning intermediate files}
}
\value{
A list of assigned prune-in common SNPs for genes.
}
\description{
EPIC imports prune-in variants following LD pruning pipeline for
common variants using PLINK.
}
\examples{
\dontrun{
out_dir = "/path/to/plink_output"
MAF <- 0.01
prunein_dir <- file.path(out_dir, "prunein", MAF)
prunein_snp_to_gene <- LD_pruning(common_snp_to_gene, prunein_dir)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
