% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{LD_pruning2}
\alias{LD_pruning2}
\title{Perform LD pruning on common variants and import prune-in variants}
\usage{
LD_pruning(common_snp_to_gene, prunein_dir)
}
\arguments{
\item{prunein_dir}{the directory of LD pruning intermediate files}

\item{common_snp_to_gene}{A list of assigned common SNPs for genes, returned from \code{divide_common_rare}.}
}
\value{
A list of assigned prune-in common SNPs for genes.
}
\description{
EPIC imports prune-in variants following LD pruning pipeline for
common variants using PLINK.
}
\examples{
\dontrun{
out_dir = "/path/to/plink_output"
trait = "LDL"
MAF = 0.01
prunein_dir <- file.path(out_dir, paste0("prunein2", "_", trait), MAF)
bfile_path <- "/path/to/bfile"
plink_path <- "/path/to/PLINK"
chr = 1
prunein_qc_snp_to_gene.chr <- LD_pruning2(prunein_snp_to_gene, plink_path, bfile_path, chr, prunein_dir)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
