% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{map_snp_to_gene}
\alias{map_snp_to_gene}
\title{Mapping SNPs to genes}
\usage{
map_snp_to_gene(gwas, anno_path = NULL)
}
\arguments{
\item{gwas}{a data frame of formatted GWAS summary statistics}

\item{anno_path}{path to the user-specified annotation dictionary returned from \code{annotate_gene}.
Default is 10kb upstream and 1.5kb downstream of each gene.}
}
\value{
A list of assigned SNPs for genes.
}
\description{
By default, a gene window with 10kb upstream and 1.5kb downstream of each gene is defined
 and SNPs are assigned to genes based on genomic positional mapping.
}
\examples{
 snp_to_gene.Demo <- map_snp_to_gene(gwas = gwas.formatted.Demo)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
