% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{plot_relevance}
\alias{plot_relevance}
\title{Plot tissue or cell type enrichment results}
\usage{
plot_relevance(enrichment_results)
}
\arguments{
\item{enrichment_results}{EPIC enrichment results, returned by \code{prioritize_relevance()}}
}
\value{
A ggplot object for enrichment visualization
}
\description{
Show barplots of inferred enrichment results for tissues or cell types
}
\examples{
\dontrun{
enrichment_plot.Demo <- plot_relevance(gtex.enrichment.joint.Demo)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
