% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{prioritize_relevance}
\alias{prioritize_relevance}
\title{Prioritizing trait-relevant tissue(s) and cell type(s)}
\usage{
prioritize_relevance(gene_pval, gene_corr.mat, gene_expr, gene.loc,
                      chrs = 1:22, verbose = FALSE)
}
\arguments{
\item{gene_pval}{a list of genes with gene-level association testing}

\item{gene_corr.mat}{a matrix of gene-gene correlations}

\item{gene_expr}{a data frame of gene expressions by tissues or cell types. For scRNA-seq data,
this can be returned from \code{process_scRNA}.}

\item{gene.loc}{a data frame of gene dictionary, only including genes of interest in each transcriptomic profile
(e.g., top 8000 highly variable genes from scRNA-seq or genes with great specificity score from GTEx v8).
This can be returned from \code{process_scRNA}.}

\item{chrs}{chromosomes. Default is \code{1:22}.}

\item{verbose}{a logical value of diagnostic messages. Default is \code{FALSE}.}
}
\value{
EPIC enrichment results
}
\description{
We devise a regression framework based on generalized least squares to identify risk loci enrichment
 to detect tissue- or cell-type-specific enrichment for a specific trait of interest.
}
\examples{
\dontrun{
gtex.enrichment.joint.Demo <- prioritize_relevance(gene_pval = combine_POET_gene_pval.Demo,
                                                   gene_corr.mat = combine_gene_corr.mat.Demo,
                                                   gene_expr = gtex.Demo, gene.loc = gtex.loc)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
