% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prunein_qc_snp_to_gene.Demo}
\alias{prunein_qc_snp_to_gene.Demo}
\title{Pre-stored prunein_qc_snp_to_gene.Demo data for demonstration purposes}
\format{Pre-computed by EPIC using pre-stored data \code{gwas.formatted.Demo} and PLINK outputs}
\usage{
prunein_qc_snp_to_gene.Demo
}
\description{
Pre-stored prunein_qc_snp_to_gene.Demo data for demonstration purposes
}
\keyword{datasets}
