% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{read_in}
\alias{read_in}
\title{GWAS summary statistics importing}
\usage{
read_in(gwas, snp_to_gene)
}
\arguments{
\item{gwas}{a data frame of formatted GWAS summary statistics}

\item{snp_to_gene}{A list of assigned SNPs for genes}
}
\value{
A list of GWAS summary statistics for genes.
}
\description{
Import GWAS summary statisics for gene
}
\examples{
prunein_gene_pval.Demo <- read_in(gwas = gwas.formatted.Demo,
                                  snp_to_gene = prunein_qc_snp_to_gene.Demo)

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
