% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{second_pruning}
\alias{second_pruning}
\title{Perform LD pruning on common variants and import prune-in variants}
\usage{
second_pruning(prunein_snp_to_gene, plink_path, bfile_path, chr, prunein_dir)
}
\arguments{
\item{prunein_snp_to_gene}{A list of assigned common SNPs for genes, returned from \code{LD_pruning}.}

\item{plink_path}{path to the PLINK toolkit}

\item{bfile_path}{path to the \code{bim} file of reference panel}

\item{chr}{numeric value of chromosome number}

\item{prunein_dir}{the directory of LD pruning intermediate files}
}
\value{
A list of assigned prune-in common SNPs for genes.
}
\description{
EPIC imports prune-in variants following LD pruning pipeline for
common variants using PLINK.
}
\examples{
\dontrun{
out_dir = "/path/to/plink_output"
trait = "LDL"
MAF = 0.01
prunein_dir <- file.path(out_dir, paste0("prunein2", "_", trait), MAF)
bfile_path <- "/path/to/bfile"
plink_path <- "/path/to/PLINK"
chr = 2
prunein_qc_snp_to_gene.chr <- second_pruning(prunein_snp_to_gene.Demo,
                                             plink_path, bfile_path, chr, prunein_dir)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
