% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{TRef.tpm}
\alias{TRef.tpm}
\title{Reference profiles obtained from single cell data of tumor infiltrating
cells.}
\format{A list of 3 elements: \describe{ \item{$refProfiles,
  $refProfiles.var}{Matrices (nGenes x nRefCells) of the normalized gene
  expression from the reference cells and the variability of this gene
  expression for each gene and each cell type} \item{$sigGenes}{A list of 80
  signature genes used to deconvolve the cell proportions} }}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
}
\usage{
TRef.tpm
}
\description{
A dataset containing the reference profiles given in TPM from various cell
types: \emph{B-}, \emph{NK-}, \emph{T-cells} and \emph{Macrophages}.
}
\details{
These were obtained from single-cell RNA-seq data from 9 donors from
the publication of Tirosh et al., 2016, Science. The samples
come from cancer metastases of melanoma (extracted from primary tumors
and non-lymphoid tissue metastases). The classification for each sample with
respect to each cell type is the one given by Tirosh et al.
}
\keyword{datasets}

