\name{validate_config}
\alias{validate_config}
\title{
Validate configuration file
}
\description{
Validate configuration file
}
\usage{
validate_config(config_file, export_env = parent.frame())
}
\arguments{

  \item{config_file}{path of configuration file}
  \item{export_env}{environment where to export variables}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
