% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_ref.Demo}
\alias{X_ref.Demo}
\title{Pre-stored X_ref.Demo data for demonstration purposes}
\format{Pre-stored by EPIC using genotypes from 1000 Genome Projects}
\usage{
X_ref.Demo
}
\description{
Pre-stored X_ref.Demo data for demonstration purposes
}
\keyword{datasets}
