\name{bp}
\alias{bp}
\title{
Mark that the numbers represent number of base pairs
}
\description{
Mark that the numbers represent number of base pairs
}
\usage{
bp(x)
}
\arguments{

  \item{x}{numeric vector. It will be convert to integers by \code{\link[base]{as.integer}}.}

}
\examples{
# There is no example
NULL

}
