% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{construct_X_super}
\alias{construct_X_super}
\title{Pseudo-SNP construction}
\usage{
construct_X_super(genotype, rare_gene_pval)
}
\arguments{
\item{genotype}{a genotype matrix}

\item{rare_gene_pval}{a list of GWAS summary statistics for rare variants}
}
\value{
A matrix of aggregated pseudo-SNP
}
\description{
We collapse genotypes of all rare variants within a gene
 to construct a pseudo-SNP
}
\examples{
\dontrun{
X_super.Demo <- construct_X_super(genotype = X_ref.Demo, rare_gene_pval = rare_gene_pval.Demo)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
