% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{package}
\name{EPIC.package}
\alias{EPIC.package}
\alias{EPIC.package-package}
\title{EPIC: a package to Estimate the Proportion of Immune and Cancer cells from
 tumor gene expression data.}
\description{
EPIC package provides the function and cell reference profiles to
estimate the proportion of immune, stromal, endothelial and cancer or other
cells from bulk gene expression data.
}
\section{EPIC functions}{

\code{\link{EPIC}} is the main function to call to estimate the
 various cells proportions from a bulk sample.
}

\section{Included datasets}{

\code{\link{BRef}}: reference profiles from circulating immune cells.

 \code{\link{TRef}}: reference profiles from tumor infiltrating non-malignant
 cells obtained from single cell data of melanoma patients.

\code{\link{melanoma_data}}: example dataset containing data from lymph nodes
 from patients with metastatic melanoma.

\code{\link{mRNA_cell_default}}: values of mRNA per cell for the main cell
 types.
}

\section{References}{

Racle, J., Jonge, K. de, Baumgaertner, P., Speiser, D.E., and
Gfeller, D. (2017). Simultaneous enumeration of cancer and immune cell
types from bulk tumor gene expression data. \emph{eLife Sciences, 6}, e26476
(\url{https://elifesciences.org/articles/26476})
}

\section{Authors}{

Julien Racle <\email{julien.racle@unil.ch}> and David Gfeller
<\email{david.gfeller@unil.ch}>.
}

