% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_fun.R
\name{EPIC}
\alias{EPIC}
\title{Estimate the proportion of immune and cancer cells.}
\usage{
EPIC(bulk, reference = NULL, mRNA_cell = NULL, mRNA_cell_sub = NULL,
  sigGenes = NULL)
}
\arguments{
\item{bulk}{A matrix (\code{nGenes} x \code{nSamples}) of the raw genes
expression from each bulk sample. This matrix needs to have rownames
telling the gene names. In principle given as raw read counts, but it
could be given in tpm as well if using corresponding reference profiles
normalized in tpm as well.}

\item{reference}{(optional): A string or a list defining the reference cells.
It can take multiple formats, either: \itemize{
  \item\code{NULL}: to use the default reference profiles and genes
    signature \code{\link{BRef}}.
  \item a char: one of \emph{"BRef"}, \emph{"BRef.tpm"} or \emph{"TRef.tpm"}
    to use the reference cells and genes signature of the corresponding
    datasets (see \code{\link{BRef}}, \code{\link{BRef.tpm}} and
    \code{\link{TRef.tpm}}).
  \item a list. When a list it should include: \describe{
    \item{\code{$refProfiles}}{a matrix (\code{nGenes} x \code{nCellTypes})
    of the reference cells genes expression (without the cancer cell type);
    the rownames needs to be defined as well as the colnames giving the
    names of each reference cell types;
    }
    \item{\code{$sigGenes}}{a character vector of the gene names to use as
      signature - sigGenes can also be given as a direct input to EPIC
      function;}
    \item{\code{$refProfiles.var}}{(optional): a matrix (\code{nGenes} x
    \code{nCellTypes}) of the variability of each gene expression for each
    cell type (if absent, we assume an identical variability for all genes
    in all cells) - it needs to have the same dimnames than refProfiles;}
    }
}}

\item{mRNA_cell}{(optional): A named numeric vector: tells (in arbitrary
units) the amount of mRNA for each of the reference cells and of the
other uncharacterized (cancer) cell. Two names are of special meaning:
\emph{"otherCells"} - used for the mRNA/cell value of the "other cells"
from the sample (i.e. the cell type that don't have any reference gene
expression profile) ; and \emph{default} - used for the mRNA/cell of the
cells from the reference profiles for which no specific value is given
in mRNA_cell (i.e. if mRNA_cell=c(Bcells=2, NKcells=2.1, otherCells=3.5,
default=1), then if the refProfiles described Bcells, NKcells and Tcells,
we would use a value of 3.5 for the "otherCells" that didn't have any
reference profile and a default value of 1 for the Tcells when computing
the cell fractions).
To note: if the data is given as raw counts, then mRNA per cell should
correspond to some weight of mRNA per cell (or to a total number of mRNA
nucleotides per cell); while if data is in tpm, then this mRNA per cell
would ideally correspond more to some number of transcripts per cell.
The default values correspond to data in raw counts.}

\item{mRNA_cell_sub}{(optional): This can be given instead of \code{mRNA_cell} (or
in addition to it). It is also a named numeric vector, used to replace
only the mRNA/cell values from some cell types (or to add values for new
cell types). The values given in mRNA_cell_sub will overwrite the default
values as well as those that might have been given by mRNA_cell.}

\item{sigGenes}{(optional): a character vector of the gene names to use as
signature for the deconvolution. In principle this is given with the
reference as the "reference$sigGenes" but if we give a value for this
input variable, it is these signature genes that will be used instead of
the ones given with the reference profile.}
}
\value{
A list of 3 matrices:\describe{
 \item{\code{mRNAProportions}}{(\code{nSamples} x (\code{nCellTypes+1})) the
   proportion of mRNA coming from all cell types with a ref profile + the
   uncharacterized other cell.}
 \item{\code{cellFractions}}{(\code{nSamples} x (\code{nCellTypes+1})) this
   gives the proportion of cells from each cell type after accounting for
   the mRNA / cell value.}
 \item{\code{fit.gof}}{(\code{nSamples} x 12) a matrix telling the quality
   for the fit of the signature genes in each sample. It tells if the
   minimization converged, and other info about this fit comparing the
   measured gene expression in the sigGenes vs predicted gene expression in
   the sigGenes.}
}
}
\description{
\code{EPIC} takes as input bulk gene expression data (RNA-seq) and returns
 the proportion of mRNA and cells composing the various samples.
}
\details{
This function uses a constrained least square minimization to estimate the
proportion of each cell type with a reference profile and another
uncharacterized cell type in bulk gene expression samples.

The names of the genes in the bulk samples, the reference samples and in the
gene signature list need to be the same format (gene symbols are used in the
predefined reference profiles). The full list of gene names don't need to be
exactly the same between the reference and bulk samples: \emph{EPIC}
will use the intersect of the genes.
}
\examples{
res1 <- EPIC(hoek_data$rawCounts)
res1$cellFractions
res2 <- EPIC(hoek_data$rawCounts, BRef)
res3 <- EPIC(bulk=hoek_data$rawCounts, reference=BRef)
res4 <- EPIC(hoek_data$rawCounts, reference="BRef")
res5 <- EPIC(hoek_data$rawCounts, mRNA_cell_sub=c(Bcells=1, otherCells=5))
res6 <- EPIC(bulk=hoek_data$rawCounts, reference="BRef.tpm")
# Various possible ways of calling EPIC function. res 1 to 4 should
# give exactly the same outputs, and the elements res1$cellFractions
# should be equal to the example predictions found in
# hoek_data$cellFractions.pred for these first 4 results.
# The values of cellFraction for res5 will be different due to the use of
# other mRNA per cell values for the B and other cells.
# And res6 will also give different results and is not advised: the reference
# BRef.tpm corresponds to tpm while the bulk was given as raw counts.

}

