% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC.R
\name{get_gene_chisq}
\alias{get_gene_chisq}
\title{Chi-square gene-level association testing and gene-gene correlation}
\usage{
get_gene_chisq(gene_pval, gene.loc, type = "POET", inter_dir)
}
\arguments{
\item{gene_pval}{a list of GWAS summary statistics for genes}

\item{gene.loc}{a data frame of gene dictionary, only including genes of interest in each transcriptomic profile
(e.g., top 8000 highly variable genes from scRNA-seq or genes with great specificity score from GTEx v8).
This can be returned from \code{process_scRNA}.}

\item{type}{type of variants included in POET estimation. This should be either 'prunein' or 'joint'.}

\item{inter_dir}{directory of intermediate output files}

\item{max.dist}{the maximum distance within which pair-wise gene-gene correlation is computed. Default is \code{5}. Unit is "Mb".}

\item{sliding_size}{size of sliding window. Default is \code{10}.}

\item{sliding_step}{the moving step of sliding-window. Default is \code{1}.}
}
\value{
A list with components
 \item{gene_POET_pval}{A list of gene-level testing}
 \item{gene_corr.mat}{A matrix of gene-gene correlations}
}
\description{
We perform gene-level chi-square association testing and calculate
 gene-gene correlations.
}
\examples{
\dontrun{
get_gene_chisq(gene_pval = prunein_gene_pval, gene.loc = gtex.loc,
               type = "POET", inter_dir = inter_dir)
}

}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
