\name{load_config}
\alias{load_config}
\title{
Load and validate configuration file
}
\description{
Load and validate configuration file
}
\usage{
load_config(config_file, export_env = parent.frame(), validate = TRUE)
}
\arguments{

  \item{config_file}{path of configuration file}
  \item{export_env}{environment where to export variables}
  \item{validate}{whether do validation}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
