% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_descr.R
\docType{data}
\name{melanoma_data}
\alias{melanoma_data}
\title{Example dataset containing data from lymph nodes from patients with
metastatic melanoma.}
\format{This is a list of 3 elements: \describe{
 \item{$counts}{(matrix of 49902 genes x 4 donors) The TPM normalized counts
   from the four donors. It has been obtained by mapping RNA-seq data to
   \emph{hg19} genome with help of \emph{RSEM}. Ensembl ID were then
   converted to gene names, and genes with duplicated entries were
   merged together by summing their counts.}
 \item{$cellFractions.obs}{(matrix of 4 donors x 6 cell types) The
   proportions of the different cell types measured by FACS (the
   "other_cells" correspond to the live cells without any marker of the
   other given cell types).}
 \item{$cellFractions.pred}{(matrix of 4 donors x 8 cell types) The
   proportions of the different cell types, as predicted by EPIC based on
   the reference profiles \code{TRef}.}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE93722}
}
\usage{
melanoma_data
}
\description{
This is the dataset measured in \href{https://elifesciences.org/articles/26476}{
 \cite{Racle et al., 2017, eLife}}. It contains
the gene expression from lymph node samples from four patients with melanoma,
and it contains also the proportions of the main immune cell types and of
melanoma cells, as measured by FACS.
}
\keyword{datasets}
